Move CourtForm and OutsideCourtForm views into view directory
