+++ client/src/components/index.js
@@ -11,7 +11,5 @@ export { default as UserCaseList } from './UserCaseList';
 export { default as MediatorPendingCaseList } from './MediatorPendingCaseList';
 export { default as MediatorActiveCaseList } from './MediatorActiveCaseList';
 export { default as MediatorCompletedCaseList } from './MediatorCompletedCaseList';
-export { default as OutsideCourtForm } from './OutsideCourtForm';
-export { default as CourtForm } from './CourtForm';
 export { default as AdminMediatorRequestList } from './AdminMediatorRequestList';
 export { default as MediatorFilter } from './MediatorFilter';
