+++ client/src/views/index.js
@@ -14,4 +14,6 @@ export { default as CaseForm } from './CaseForm'
 export { default as UserCaseShow } from './UserCaseShow';
 export { default as MediatorCasesShow } from './MediatorCasesShow';
 export { default as AdminHome } from './AdminHome';
-export { default as AuthCallback} from './AuthCallback';
+export { default as AuthCallback } from './AuthCallback';
+export { default as CourtForm } from './CourtForm';
+export { default as OutsideCourtForm } from './CourtForm';
